refactor: moved interface ISignIn and ISignUP  to page files
