--- src/types/interfaces/ISignInForm.tsx
@@ -1,4 +1,0 @@
-export interface ISignInForm {
-  email: string;
-  password: string;
-}
