--- src/types/interfaces/ISignUpForm.tsx
@@ -1,6 +1,0 @@
-export interface ISignUpForm {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-}
